Bf 3 (#38)
* QOL verify and attend

Use of embeds for verify and attend. Also improved text.

* QOL startm

Text improvements for ticket system.

* U and QOL Permissions

Permissions now work! We use update instead of set permissions so that they hold the category's permissions!

* BF Ticket Permissions

Fixed some bugs with ticket permissions. It all works now!

* U Keep Archive Bottom

We now keep the archive on the very bottom!

* BF startm

Fixed a bug realted to undeclared variable due to previouse PR.

* BF Permission REMOVED HIDE/UNHIDE

The hide /unhide functionality is bringing a lot of issues with permissions and  category locking and unlocking so I have decided that the best option is just to remove it. The code is still present but it is disabled and shouldnt be used!

* BF Passowrd Stamp and Workshop QOL

Bug fixes for password stamp, changes TA assistance one-liner timer to 20 seconds. Embed improvements for polls.