Custom command class refactoring (#69)
* Custom Command Class

Created the custom command class. It extends the Command class from discord.js-commando. It needs a Client, CommandInfo and CommandPermissionInfo. Tested it with clear chat and it works!

* Custom Command Improvements

Changed customCommand name to PermissionCommand. Added more documentation and checks for required child functions.

* Permission Command use in a-utility

All relevant commands in a_utility now extend the PermissionCommand class.

* PermissionCommand use in utility and verification

Verificaiton commands and the ask command on utility extend the PermissionCommand class.

* PermissionComand use in a_start_commands

All a_start_commands commands extend PermissionCommand.

* PermissionCommand use in a_boothing

All commands in the a_boothing group extend PermissionCommand. These commands use channel name not ID. Will not make a change since this will change with the boothing upgrades.

* BF admin console channel

DiscordServices exports the admin console channel ID, it is used for start-mentor-room.js command.

* Permission Command file name change

Changed the PermissionCommand file name to permission-command.js

* ActivityCommand class

Uncomplete ActivityCommand class declaration. Might not use it so its not complete.

* Prompt class

Created the Prompt class. It contains static varaibles usefull to prompt the user for input. Currently only has one, it prompts for a text message.

* Better stamp role management

DiscordServices now has a Collection with the stamp number as key and stamp role ID as value. Eaiser to use!

* Activity and ActivityManager Creation

Created the Activity class and its manager. The activity represents a discord activity, it and its manager contains common code for all of the commands. They are still not used for the commands.

* BF

Bug fixes for promt and activity manager classes.

* BF activity command and callback implementation of activity class

Bug fixes on activity, finished implementing the activitycommand class and it is implemented on the callback command. It works!

* activity-command activity check

Checks to make sure there is an activity to be used for the command. Calling the command from outside the activity console will result in the command not being called and a warning being sent.

* added number prompt

Added a second static method, it prompts for and returns a number.

* BF on activity, activity command and permission command

small bug fixes and improvements on the activity, activity command and permission command classes.

* Use of OOP for all activity commands

Use of the activity and activityManager classes, as well as the acticityCommand class for all the activity commands.

* It works!

No double activities! Yey!