Firebase schema improvements (#71)
* Announcements Update

Announcements are received from the nwHacks2021 doc.

* Verify Firebase Updates

Udpated two helper functions to verify users. One is for hackers, the other is for everyone else. The public verify uses these two. Updated verify command to use this correctly. Hackers are checked from CMS data.

* BF Discord Services

Error text BF and guest role BF.

* Removed Discord data on Firebase

We are not adding isVerified, isAttending or discordID to firebase at the moment. We would have to sign in as admin. We will just use permission to keep people from attending/verifing twice.