Team and team member options
Added options for the team leader and team members to leave the team
both before a full team is created and after the team is created. After
every change, the system will try to find a team for the new team size.
Added a isComplete helper and a hasBeenComplete to Team class. Roulette
has some new await reactions.