Mentor ticket bugs (#102)
* Raffle Bug Fix

Added a end of comment line.

* Restrict One Ticket at a time

Added a collection of users for users who are submitting a ticket. The collector filter will not let users who are in this list to submit a ticket.
Users are added as soon as they enter the collector.

* Remove user from currentlySubmitting list

Forgot to remove the user once they submit the ticket. It is now removed
after the user submits the ticket information.

* Emoji on Ticket Card Fix

Fixed an emoji not showing up properly on the metnor ticket card by calling
the toString() function on the emoji.

* Fix enter ticket permissions

When users entered the ticket, SEND_MESSAGES was not given, and when they
leave the ticket, they loose SEND_MESSAGES. Added SEND_MESSAGES to the
permission variable.

* Ticket Number Fix

Ticket number used for ticket and channel was a command variable. Now we
give each ticketInfo class its own ticket id (ticket number). This id is
used for everything. Changed createTicketChannel helper function to
accept a ticketInfo to use. Very important to have the correct number for
the channel.