Team roulette improvements (#103)
* Added a sendEmbedToUser helper function

This function needs a user/member, some embed options and if the message
should be deleted. It makes sure the embed info is filled.

* Added an empty string check to sendEmbedToUser

Added an empty string check, if failed an error is thrown.

* Team roulette DMs Changes

All team roulette DMs are now embeds with a longer timer. This will ensure
users know what is going on. Also any succesful message is not deleted.
Uses the new sendEmbedToUser helper function.

* Created the Team classes

The Team class represents a real life team with members. The primary use
of this class at the moment is with the team roulette, but it could be
used for other functions in the future.

* Added addMember function and team leader property

The addMember function will NOT add any dupllicate members and will
automatically detect the first member and make him the team leader.

* More Team functions

Added a remove team member function as well as a team size funciton. The
merge team function returns the team.

* Smarter merge team function

The merge team function will perserve the team with a text channel if
there is any.

* Team new member introduction

Bot sends an introduction to the text channel when new members are added.

* Null Safety

Added some null safety to the class functions when they check for a
text channel.

* Use of Team class for roulette and remove function

The roulette functionality uses the Team class to better organize all
the teams. Furthermore, the use of this class facilitates the removal
of members and will faciliate further changes.

* Documentation fix on sendEmbedToMember

* Team and team member options

Added options for the team leader and team members to leave the team
both before a full team is created and after the team is created. After
every change, the system will try to find a team for the new team size.
Added a isComplete helper and a hasBeenComplete to Team class. Roulette
has some new await reactions.

* Added custom channel and category

Command can be ran anywhwere, it will ask if a channel is needed, if so
then create it, else ask for it. Channel is cleaned and added to the
black list.

* Only available on the admin channel

* Refactor channel creation or prompt

Refactored out the channel prompt and/or creation to a helper funciton.

* Added admin view and Team toString

Admins can request a list of all the teams waiting on the team roulette.
This functionality uses a new toString of the Team class. To request
the list, admins must react to a team roulette console found on the
admin console.

* Add field options to sendEmbedToMember helper

Made this possible with a new typedef FieldInfo, has a title and
description. Function will ask for a list of FieldInfo, iterates and
adds the info as fields to the embed.

* json file for debug settings

* Fixed awaitRection Bug

The awaitRection call had no max, so it was not calling the listener
after the first and only emoji. Added a max of 1 and it works.