Password stamp improvements (#109)
* Generalized wording to accept all activities, not only booths.

* Let hacker sknow about the password requirements

* Documentation improvements for sendMessageToMember

* Use of discordServices helpers to use embeds for DMs

* Better collector filtering with Collection

For the stamp collector, instead of using a list we use a collection.
The collection holds user's IDs and usernames. Any user already in this
collection will not be able to activate the collector.

* Safer command by adding permissions

The command now extends PermissionCommand and it must be called by
staff and only on the admin console. Also prompts the staff for a
channel to send the stamp collector to.

* Remove unecessary checks due to uprgade to PermissionCommand

* Use Prompt class to prompt activity name and password

* Spelling fix

* Bug fix on prompt