PermissionCommand class bug fix with new discordServices
Commands are constructed when they are loaded, not every time the are
called. This caused the permissions to use the discordServices file
before it is populated and led to wrong permissions. Now we use
constants that are then used to grab the correct role or channel from
discordServices. This will also make it easier to use the BotGuild class
in the future.