Bug fix empty channel/role/mention
It would check the return of the message properties to null, but an
empty collection would be returned and so that check was useless. It
now checks to see if there is an element by grabbing the first one, if
this is not the case it asks the user to try again.