Botguild class creation Updated Bug Fixes (#156)
* Create bot-guild.js

Created the class with a constructor.

* Added minimum information setup function and static admin channel creation

* Added the verificaiton setup function. Fixed permission bug.

* Added the attendance setup function.

* Added the announcement setup function, not complete, need to add firebase code.

* Added the stamp setup function, creates stamps role for us!

* Added the report setup function.

* Async bug fixes

* Added the ask command setup function

* Added group enable after setup success

* Use of BotGuild class in init-bot commands

The init-bot command uses the BotGuild class instead of discordServices.
Added the ability to use pre made stamps and verification channels for
better use with saved data. Fixed some async bugs.

* Update prompt.js

Made optional params optional on JSDoc

* Update report.js

Bug fix due to async error.

* Update dev_config.json

Updated to use new role and channel IDs.

* Install Mongo and Mongoose

* Create botGuildDBObject.js

Created the BotGuild schema and Model for mongoose.

* Create mongoUtil.js

Created module to connect to mongodb and mongoose.

* Update index.js

Fixed but log errors, re-arranged imports and cleaned out firebase calls.

* Update index.js

Fixed but log errors and cleaned out imports and firebase calls.

* Update bot-guild.js

Removed properties, changed role ID to be string, fixed async bugs, removed guildID param, fixed MessageEmbed import error, moved permissions to be static params.

* Update init-bot.js

Async fixes and use of Mongoose BotGuild model!

* Use of Mongoose Object instead of DiscordServices

Changed all discordServices RoleIDs usage with mongoose.

* Update init-bot.js

Checks if botGuild is complete and returns if that is the case. Will set isSetUpComplete to true after using json data.

* Update init-bot.js

Give the user setting up the bot the admin role after its creation.

* Update index.js

Clean up on message action and fix bug with it being true always. Use of botGuild to find blacklist.

* Update index.js

Use of botGuild for verification and new member greeting.

* Update package-lock.json

Updated mongodb package version.

* Update discord-services.js

Removed all the properties that are not used any more with the introduction of botGuild. Changed the use of such parameters in the functions with botGuild calls.

* Update botGuildDBObject.js

Spelling fix with channelIDs property name.

* Update activity-command.js

Activity command passes the botGuild to its activityCommand() function for commands to use.

* Activity Commands Update use of botGuild

Updated all the ActivityCommand commands to accept the botGuild as its first parameter for them to be usable. Removed the use of discordServices properties and added the use of botGuild.

* Update activity.js

Use of botGuild for properties instead of discordServices.

* Update cave.js

Use of botGuild for properties instead of discordServices. Added a property with the botGuild.

* Update team-formation.js

Use of botGuild for properties instead of discordServices.

* Update bot-guild.js

But fix for commented out code.

* Update permission-command.js

PermissionCommand passes a botGuild to the runCommand() method for the commands to use.

* Update all PermissionCommand sub classes to use botGuild

Use of botGuild for properties instead of discordServices.

* Update e-room-directory.js

Use of botGuild for properties instead of discordServices.

* Boothing Commands Bug Fixes

Fixed bug with wrong permission flags and use of botGuild.

* Verification NOT WORKING

Verify and attend do not work since they are through DM there is not way to get guild ID. They must be fixed and used in case the other options are not working.

* Update activity-manager.js

Use of botGuild for properties instead of discordServices.

* Remove the isAdminConsole method

This discordServices method is used twice, and its better to just
use the local botGuild. So we just removed it.

* Use of cave botGuild in ticket createCategory

* Ticket Content and Password fixes

* Removing mentor and sponsor role use from ask and e-room

* Use of previously fetched botGuild in index and report

* Use of command botGuild for activity manager functions

* Activity save fixes and use of enabled

* Import bug fixes for index.js

* Replace discordServices with botGuild

* Add verificationiRoles map to the verificaiton property of botGuild

* Verification Types in InitBot commands

Initbot command will ask the user for verification types. The dev_config
file will also add two types to the bot. Types are always in lower case!