Improved guild command enable/disable
Added a ask property with isEnabled to the botGuild class. With this
I added a function to the botGuild class that will set its commands
to enabled/disabled by looking at its data, you need to pass it a
client. This makes sure guilds keep the same settings even after a
bot restart.