Stamp command block (#166)
* Creation of Stamps Command Group

The stamps command group holds all the stamp related commands. This
group can only be turned on if the stamps are set up during bot-init.

* Added failsafe to parseRole function, throws error

* added logger message when parseRole throws error

* Added guard to activity and workshop consoles against stamp use when disabled

* Bug fix with command enable/disable

When the bot restarts, the enabled/disabled settings are lost. Added
a guard when bot starts to set commands to disabled if a guild is new or
 if the guild is not set up.

* Improved guild command enable/disable

Added a ask property with isEnabled to the botGuild class. With this
I added a function to the botGuild class that will set its commands
to enabled/disabled by looking at its data, you need to pass it a
client. This makes sure guilds keep the same settings even after a
bot restart.