Changed mentor shuffle to role shuffle (#163)
The Activity Manager function expects a role, it will shuffle users in
the general voice channel with that role down to the other voice
channels. Uses the activity's voiceChannel properties, so it will only
work with channels created from the admin console. shuffle-mentor command
changed to shuffle-role, it prompts for a role to pass on to
Activity Manager function. Added logging to this command.