Activity Code Restructure
Code was moved around to follow the following format:
- Constructor
- init
- features
- class functions
- feature callbacks

There is a comment between features and class functions. Moved the
addDefaultFeatures() call from constructor to the init() function.
Channels property has safeChannels, this channels can't be removed by
users. AddChannelHelper has an option to add the created channel to
this list.