Activities rework (#174)
* Fixed commands being disabled, had to read commandGroup ID not name

* Removed debug logs

* Improved error logging

* Bug fixes on logging statements

* Stamps small fix, did not set commands enabled after setting up stamps

* Added logging to check issues, found none, left logging

* Workshop and Activity rework

Activity is a single principle class. Workshop class extends the
activity class to add workshop functionality. !init-workshop creates
a new workshop. No activity console yet. Moved all activity related
classes to /classes/activities/ folder. Old activities class still in
use, file called old-activities.js.

* JSDoc activity class documentation

* Activity static function for role prompts

Added a static function to the Activity class to prompt the user for
the roles that can participate in the activity as it will be used by
all other subclasses when instantiating themselves. Using it on
new-activity command.

* Created the CoffeeChats classes

The coffeeChats class extends the activity class, it sets up the
coffee chats activity, it took away the group shuffle from the
activity manager. Uses an collection to store the groups.

* More functions for the coffee chats

When adding voice channels, the number of groups increases as well.
Function to reset the teams. Added a timer to prompt stating the
activity is full.

* Bug fix on logger

* init() returns self

* Rename commands to new- instead of init-. JSDoc bug fix

* new-workshop use of activity static prompt for participant roles

* Added archive category to botGuild

The archive category will be used by activities, it is created by default with the botGuild setUp() function. In the future we could not create it until needed.

* Added features property to Activity class

The features collection holds all the features of the activity, these features hold a name, description, emoji and callback. Features are presented in the activity console. When the emoji is reacted, the callback is triggered.

* Added the sendAdminConsole function to the activity class, called in init().

* Async bug fix

* Bug fixes on archive category creation

* Added features to activity

Every activity has a list of features, these features are added to the
admin console and have an emoji and a callback. The callback is triggered
when the emoji is reacted. Subclasses can add features, they must call
the super after adding the features. Activity constructor calls the
function.

* Activity delete removes the console message

* Bug fixes for all activities

* Added Voice Callback feature to activity

The voice callback will move all members in activity's voice channels to
one voice channel decided by the user. For this, added a discordServices
function that shows the user a list of channels, the user can then choose
a channel using numbers. Also, feature callbacks will get the user who
triggered the feature.

* Const used to set server loggers to log to console

* Added shuffle and role shuffle to activity

These two features will shuffle all members or those with a role
around the activity's voice channels. Role shuffle uses shuffle, shuffle
accepts a filter.

* Added logging to the admin console collector

* Moved stamp functions to stamps-manager class.

* Added distribute stamp feature to activity and removed activity manager file

* Bug fix on distribute stamp

* Added Rule Validation and Archive fix

Rule validation is what make among us was but now generalized to any
kind of activity. It locks the activity behind an emoji reaction,
the user must read the rules and react to gain access. Uses individual
permissions to work. Also fixed Archive feature.

* Clean out New Activity and commands

!new-activity now uses the new activity class. Removed all previous
activity commands that will not be used. Changed activity features embed
to put emoji in field title.

* Coffe chats init use new chooseChannel function

* Added a reaction picker, given options user reacts to choose

* Improved add/remove channels

Users can now add and remove both voice and text channels. To add a
channel, user must react, the decide between voice or text using new
reaction chooser, then give name. Workshops also ask if the channel
is for TA. To remove, user reacts and then chooses from a list by
using the discordServices channel chooser.

* Moved channel chooser to prompt class

* JSDoc bug fix

* Import fixes

* Bug fixes on chooseChannel

* Activity Code Restructure

Code was moved around to follow the following format:
- Constructor
- init
- features
- class functions
- feature callbacks

There is a comment between features and class functions. Moved the
addDefaultFeatures() call from constructor to the init() function.
Channels property has safeChannels, this channels can't be removed by
users. AddChannelHelper has an option to add the created channel to
this list.

* Coffee Chats Code Restructure

Following the activity restructure. Added the features to the features
list with a addDefaultFeatures(). Added a add team slot feature. Regular
add channel does not add team slots. Bug fixes on groupShuffle().

* Workshop Code Restructure

Followed new structure in activity class. Polls are added dynamically to
the features list in addDefaultFeatures(). Had to add an emojiName
property to the pollInfo typedef. Polling will send it to general text
or prompt for a channel if not available.

* Activity addDefaultFeatures uses workshop and coffee chats better format using one list and then foreach

* Removed ActivityCommand as it is not used any more

* More merge fixes

* More merge fixes pt 2