Created a new class room
The new Room class represents a category with voice and text channels.
The Room class can add, remove channels, users and roles able to view
the category. Lock the room, delete it and archive it. An activity
lives in a room. Workshops and coffee chats have also been updated to
use the room class given from the activity class.